feat: 당첨 로또에서 당첨 결과를 구하는 기능 구현
