+++ README.md
@@ -29,7 +29,7 @@
 ### 당첨 로또
 - [x] 6개의 당첨 번호와 보너스 번호를 갖는다.
 - [x] 당첨 번호와 보너스 번호는 중복되지 않는다.
-- [x] 로또 번호를 비교하여 로또 등수를 구한다.
+- [x] 구입한 로또와 비교하여 당첨 결과를 구한다.
 
 ### 로또 발매기
 - 자동 로또 발매기
@@ -43,6 +43,3 @@
 ### 로또 당첨 결과
 - [x] 등수별 당첨 횟수를 갖는다.
 - [x] 당첨 수익률을 계산한다.
-
-### 로또 당첨 계산기
-- [x] 당첨 로또와 구입한 로또를 비교하여 당첨 결과를 구한다.
