+++ src/test/kotlin/lotto/domain/model/LottoResultTest.kt
@@ -1,22 +1,11 @@
 package lotto.domain.model
 
-import lotto.domain.value.LottoNumber
 import lotto.domain.value.PurchaseAmount
 import lotto.enums.Rank
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
 class LottoResultTest {
-    private lateinit var winningLotto: WinningLotto
-
-    @BeforeEach
-    fun setUp() {
-        val winningNumbers = Lotto.of(1, 2, 3, 4, 5, 6)
-        val bonusNumber = LottoNumber.from(45)
-        winningLotto = WinningLotto(winningNumbers, bonusNumber)
-    }
-
     @Test
     fun `14장의 로또 중 1장이 5등일 때 당첨 수익률을 계산한다`() {
         // give
