+++ src/test/kotlin/lotto/domain/model/WinningLottoTest.kt
@@ -3,23 +3,16 @@ package lotto.domain.model
 import lotto.domain.value.LottoNumber
 import lotto.enums.Rank
 import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 import org.junit.jupiter.api.assertThrows
 
 class WinningLottoTest {
-    private lateinit var winningNumbers: Lotto
-    private val bonusNumber = LottoNumber.from(45)
-
-    @BeforeEach
-    fun setUp() {
-        winningNumbers = Lotto.of(1, 2, 3, 4, 5, 6)
-    }
-
     @Test
     fun `당첨 번호와 보너스 번호는 중복되지 않는다`() {
         assertDoesNotThrow {
+            val winningNumbers = Lotto.of(1, 2, 3, 4, 5, 6)
+            val bonusNumber = LottoNumber.from(45)
             WinningLotto(winningNumbers, bonusNumber)
         }
     }
@@ -27,50 +20,42 @@ class WinningLottoTest {
     @Test
     fun `당첨 번호와 보너스 번호가 중복되면 예외가 발생한다`() {
         assertThrows<IllegalArgumentException> {
+            val winningNumbers = Lotto.of(1, 2, 3, 4, 5, 6)
             val bonusNumber = LottoNumber.from(1)
             WinningLotto(winningNumbers, bonusNumber)
         }
     }
 
     @Test
-    fun `일치하는 번호가 6개이면 등수는 1등이다`() {
-        val lotto = Lotto.of(1, 2, 3, 4, 5, 6)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.FIRST)
+    fun `1등 당첨 로또가 1개이면 1등 당첨 횟수는 1이다`() {
+        // given
+        val winningLotto = WinningLotto(Lotto.of(1, 2, 3, 4, 5, 6), LottoNumber.from(45))
+        val firstRankLotto = Lotto.of(1, 2, 3, 4, 5, 6)
+        val lottos = listOf(firstRankLotto)
+
+        // when
+        val lottoResult = winningLotto.getLottoResult(lottos)
+
+        // then
+        val actual = lottoResult.winningStats[Rank.FIRST]
+        val expected = 1
+        assertThat(actual).isEqualTo(expected)
     }
 
     @Test
-    fun `일치하는 번호가 5개이고 보너스 번호가 일치하면 등수는 2등이다`() {
-        val lotto = Lotto.of(1, 2, 3, 4, 5, 45)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.SECOND)
-    }
+    fun `당첨되지 않은 로또가 1개이면 미당첨 횟수는 1이다`() {
+        // given
+        val winningLotto = WinningLotto(Lotto.of(1, 2, 3, 4, 5, 6), LottoNumber.from(45))
+        val firstRankLotto = Lotto.of(1, 2, 3, 4, 5, 6)
+        val missRankLotto = Lotto.of(11, 12, 13, 14, 15, 16)
+        val lottos = listOf(firstRankLotto, missRankLotto)
 
-    @Test
-    fun `일치하는 번호가 5개이고 보너스 번호가 일치하지 않으면 등수는 3등이다`() {
-        val lotto = Lotto.of(2, 3, 4, 5, 6, 7)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.THIRD)
-    }
+        // when
+        val lottoResult = winningLotto.getLottoResult(lottos)
 
-    @Test
-    fun `일치하는 번호가 4개이면 등수는 4등이다`() {
-        val lotto = Lotto.of(3, 4, 5, 6, 7, 8)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.FOURTH)
-    }
-
-    @Test
-    fun `일치하는 번호가 3개이면 등수는 5등이다`() {
-        val lotto = Lotto.of(4, 5, 6, 7, 8, 9)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.FIFTH)
-    }
-
-    @Test
-    fun `일치하는 번호가 2개 이하이면 등수는 MISS이다`() {
-        val lotto = Lotto.of(5, 6, 7, 8, 9, 10)
-        val winningLotto = WinningLotto(winningNumbers, bonusNumber)
-        assertThat(winningLotto.getRank(lotto)).isEqualTo(Rank.MISS)
+        // then
+        val actual = lottoResult.winningStats[Rank.MISS]
+        val expected = 1
+        assertThat(actual).isEqualTo(expected)
     }
 }
